% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mapiso-package}
\alias{mapiso-package}
\title{Package description}
\description{
Regularly spaced grids containing continuous data are
transformed into contour polygons. A grid can be defined by a
data.frame (x, y, value), an \code{sf} object or a \code{terra} SpatRaster.
}
