% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapiso.R
\name{mapiso}
\alias{mapiso}
\title{Title}
\usage{
mapiso(x, var, breaks, nbreaks = 8, mask, coords, crs)
}
\arguments{
\item{x}{a data.frame, an sf data.frame or a SpatRaster.}

\item{var}{df and sf only}

\item{breaks}{list of break values (default to equal interval)}

\item{nbreaks}{nclass number of classes}

\item{mask}{sf object}

\item{coords}{df only, vector of names (x and y)}

\item{crs}{df only epsg or other code}
}
\value{
and sf object of iso polygones
}
\description{
Title
}
\examples{
# terra
if(require(terra, quietly = TRUE)){
  r <- rast(system.file("tif/elevation.tif", package = "mapiso"))
  isor <- mapiso(x = r)
  plot(r)
  library(sf)
  plot(st_geometry(isor), add = TRUE, col = NA)
}

# sf, using a mask
s <- st_read(system.file("gpkg/elevation.gpkg", package = "mapiso"),
             layer = "elevation", quiet = TRUE)
m <- st_read(system.file("gpkg/elevation.gpkg", package = "mapiso"),
             layer = "com", quiet = TRUE)
isos <- mapiso(x = s, var = "elevation",
               mask = m)
plot(isos)

# data.frame, using user breaks values
d <- read.csv(system.file("csv/elevation.csv", package = "mapiso"))
bks <-c(98,100, 150, 200, 250, 300, 350, 400, 412.6)
isod <- mapiso(x = d, var = 'elevation',
               breaks = bks, coords = c('x', 'y'), crs = 'epsg:2154')
plot(isod)
if(require(mapsf, quietly = TRUE)){
  mf_map(isod, "isomin", "choro", breaks = bks, leg_title = "Elevation")
}
}
