% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapiso.R
\name{mapiso}
\alias{mapiso}
\title{Title}
\usage{
mapiso(x, var, breaks, nbreaks = 8, mask, coords, crs)
}
\arguments{
\item{x}{a data.frame, an sf data.frame or a SpatRaster.}

\item{var}{df and sf only}

\item{breaks}{list of break values (default to equal interval)}

\item{nbreaks}{nclass number of classes}

\item{mask}{sf object}

\item{coords}{df only, vector of names (x and y)}

\item{crs}{df only epsg or other code}
}
\value{
and sf object of iso polygones
}
\description{
Title
}
\examples{
plot(1:10)
}
